incapsulated components, separate entrypoint for dev
