+++ flutter_material/.dart_tool/flutter_build/9b0beb5558d8b74808df34a9b5eba536/flutter_assets.d
@@ -1,1 +1,1 @@
- /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/MaterialIcons-Regular.otf /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/shaders/ink_sparkle.frag /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/AssetManifest.json /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/FontManifest.json /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/NOTICES.Z:  /Users/andrew_m/Projects/flutter_test_project/flutter_material/pubspec.yaml /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/assets/CupertinoIcons.ttf /Users/andrew_m/Projects/Flutter/bin/cache/artifacts/material_fonts/MaterialIcons-Regular.otf /Users/andrew_m/Projects/Flutter/packages/flutter/lib/src/material/shaders/ink_sparkle.frag /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/async-2.9.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.12/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.5/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.8.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.2/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.12/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/LICENSE /Users/andrew_m/Projects/Flutter/bin/cache/pkg/sky_engine/LICENSE /Users/andrew_m/Projects/Flutter/packages/flutter/LICENSE
+ /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_2.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_3.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_1.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/avatar.jpeg /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/MaterialIcons-Regular.otf /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/shaders/ink_sparkle.frag /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/AssetManifest.json /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/FontManifest.json /Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/NOTICES.Z:  /Users/andrew_m/Projects/flutter_test_project/flutter_material/pubspec.yaml /Users/andrew_m/Projects/flutter_test_project/flutter_material/assets/screen_2.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/assets/screen_3.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/assets/screen_1.png /Users/andrew_m/Projects/flutter_test_project/flutter_material/assets/avatar.jpeg /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/assets/CupertinoIcons.ttf /Users/andrew_m/Projects/Flutter/bin/cache/artifacts/material_fonts/MaterialIcons-Regular.otf /Users/andrew_m/Projects/Flutter/packages/flutter/lib/src/material/shaders/ink_sparkle.frag /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/async-2.9.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.12/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.5/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.8.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.2/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.1/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.12/LICENSE /Users/andrew_m/Projects/Flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/LICENSE /Users/andrew_m/Projects/Flutter/bin/cache/pkg/sky_engine/LICENSE /Users/andrew_m/Projects/Flutter/packages/flutter/LICENSE /Users/andrew_m/Projects/flutter_test_project/flutter_material/DOES_NOT_EXIST_RERUN_FOR_WILDCARD657093782
