+++ flutter_material/.dart_tool/flutter_build/9b0beb5558d8b74808df34a9b5eba536/outputs.json
@@ -1,1 +1,1 @@
-["/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/Flutter.framework/Flutter","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/vm_snapshot_data","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/isolate_snapshot_data","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/kernel_blob.bin","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/App","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/Info.plist","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/MaterialIcons-Regular.otf","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/shaders/ink_sparkle.frag","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/AssetManifest.json","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/FontManifest.json","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/NOTICES.Z"]
+["/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/Flutter.framework/Flutter","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/vm_snapshot_data","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/isolate_snapshot_data","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/kernel_blob.bin","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/App","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/Info.plist","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_2.png","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_3.png","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/screen_1.png","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/assets/avatar.jpeg","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/MaterialIcons-Regular.otf","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/shaders/ink_sparkle.frag","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/AssetManifest.json","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/FontManifest.json","/Users/andrew_m/Projects/flutter_test_project/flutter_material/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/NOTICES.Z"]
